+++ CyclicBuffer.hpp
@@ -30,32 +30,32 @@
 
 template <typename T>
 class CyclicBuffer {
-	size_t pos;
+	uint32_t pos;
 protected:
-	size_t mask, count, alloc_size;
+	uint32_t mask, count, alloc_size;
 public:
 	T *storage, *data;
 
-	forceinline size_t getPos() const {return pos;}
-	forceinline size_t getMask() const {return mask;}
+	forceinline uint32_t getPos() const {return pos;}
+	forceinline uint32_t getMask() const {return mask;}
 	forceinline T* getData() { return data; }
 
-	inline size_t prev(size_t pos, size_t count) const {
+	inline uint32_t prev(uint32_t pos, uint32_t count) const {
 		// Relies on integer underflow behaviour -> infinity. Works since pow 2 size.
 		return (pos - count) & mask;
 	}
 
-	inline size_t next(size_t pos, size_t count) const {
+	inline uint32_t next(uint32_t pos, uint32_t count) const {
 		return (pos + count) & mask;
 	}
 
 	// Maximum size.
-	forceinline size_t getSize() const {
+	forceinline uint32_t getSize() const {
 		return mask + 1;
 	}
 
 	// Current nubmer of items inside of the buffer.
-	forceinline size_t getCount() const {
+	forceinline uint32_t getCount() const {
 		return count;
 	}
 
@@ -77,21 +77,21 @@ public:
 		++count;
 	}
 
-	forceinline T& operator [] (size_t offset) {
+	forceinline T& operator [] (uint32_t offset) {
 		return data[offset & mask];
 	}
 
-	forceinline T operator [] (size_t offset) const {
+	forceinline T operator [] (uint32_t offset) const {
 		return data[offset & mask];
 	}
 
-	forceinline T operator () (size_t offset) const {
+	forceinline T operator () (uint32_t offset) const {
 		return data[offset];
 	}
 
 	virtual void release() {
 		pos = alloc_size = 0;
-		mask = static_cast<size_t>(-1);
+		mask = static_cast<uint32_t>(-1);
 		delete [] storage;
 		storage = data = nullptr;
 	}
@@ -101,22 +101,22 @@ public:
 	}
 
 	// Can be used for LZ77.
-	void copyStartToEndOfBuffer(size_t count) {
-		size_t size = getSize();
-		for (size_t i = 0;i < count;++i) {
+	void copyStartToEndOfBuffer(uint32_t count) {
+		uint32_t size = getSize();
+		for (uint32_t i = 0;i < count;++i) {
 			data[size + i] = data[i];
 		}
 	}
 
 	// Can be used for LZ77.
-	void copyEndToStartOfBuffer(size_t count) {
-		size_t size = getSize();
-		for (size_t i = 0;i < count;++i) {
+	void copyEndToStartOfBuffer(uint32_t count) {
+		uint32_t size = getSize();
+		for (uint32_t i = 0;i < count;++i) {
 			storage[i] = storage[i + size];
 		}
 	}
 
-	void resize(size_t newSize, size_t padding = sizeof(size_t)) {
+	void resize(uint32_t newSize, uint32_t padding = sizeof(uint32_t)) {
 		// Ensure power of 2.
 		assert((newSize & newSize - 1) == 0);
 		delete [] storage;
